+++ sycl/test/check_device_code/fpga_mem_constructor.cpp
@@ -22,7 +22,7 @@ public:
     secret = 0;
 
     for (int i = 0; i < val; i++) {
-      secret += std::sqrt(i);
+      secret += (int)sqrt((float)i);
     }
   }
 };
